back-end/requirements.txt
pydantic
pyhumps
fastapi